# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nanumlab SEO Traffic Manager API",
    "description": "API endpoints for monitoring Naver Shopping product ranks and controlling automated traffic/SEO scripts.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/product/rank": {
      "post": {
        "summary": "Query Specific Product Rank",
        "description": "Queries the current rank for a specific keyword and product ID on Naver Shopping.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyword": { "type": "string" },
                  "product_id": { "type": "string" }
                },
                "required": ["keyword", "product_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "rank": { "type": "integer" },
                    "display": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}